% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_materials_path}
\alias{get_materials_path}
\title{Get path to learning materials}
\usage{
get_materials_path()
}
\description{
This package includes lecture slides and other materials. This function shows you where to find them.
}
