% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keitt.ssi.2019-package.R
\docType{package}
\name{keitt.ssi.2019-package}
\alias{keitt.ssi.2019}
\alias{keitt.ssi.2019-package}
\title{keitt.ssi.2019: Learning materials for Geospatial Data Analysis in R, part of the UT Summer Statistics Institute}
\description{
Contains documents for learning geospatial data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thk686/keitt.ssi.2019}
  \item Report bugs at \url{https://github.com/thk686/keitt.ssi.2019/issues}
}

}
\author{
\strong{Maintainer}: Timothy Keitt \email{tkeitt@utexas.edu}

}
\keyword{internal}
