% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keitt.ssi.2019-package.R
\docType{package}
\name{keitt.ssi.2019-package}
\alias{keitt.ssi.2019}
\alias{keitt.ssi.2019-package}
\title{keitt.ssi.2019: Learning materials for Geospatial Data Analysis in R, part of the UT Summer Statistics Institute}
\description{
Contains documents for learning geospatial data analysis.
}
\author{
\strong{Maintainer}: Timothy Keitt \email{tkeitt@utexas.edu}

}
\keyword{internal}
